refactored tests to work both in python2 and python3

+++ run_tests.sh
@@ -15,14 +15,21 @@ do
     echo "Testing with"
     echo "python version: ${python_version}"
     echo "django version: ${django_version}"
-    docker-compose build \
+
+
+    if [[ ${python_version} =~ ^2. ]] && [[ ${django_version} =~ ^2. ]]; then
+        echo "django version 2 not supported in python 3 skipping this test"
+    else
+        docker-compose build \
         --build-arg PYTHON_IMAGE_TAG=${python_version} \
         --build-arg DJANGO_VERSION=${django_version} app
-    for host in "${supported_db_host[@]}"
-    do
-        sudo find . -name \*.pyc -delete
-        echo "db host: ${host}"
-        docker-compose run -e SCYLLA_HOST=${host} app python manage.py test
-    done
+
+        for host in "${supported_db_host[@]}"
+        do
+            sudo find . -name \*.pyc -delete
+            echo "db host: ${host}"
+            docker-compose run -e SCYLLA_HOST=${host} app python manage.py test
+        done
+    fi
   done
 done
